Add item test working; moving code around
